moved all auth styling into styles.js
