+++ App.js
@@ -3,15 +3,10 @@ import React, {useState, useEffect} from 'react';
 //navigation imports
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-
 import auth from '@react-native-firebase/auth';
-
-import { SafeAreaProvider } from 'react-native-safe-area-context';
-
 //custom screens
 import AuthStack from './screens/authentication/authStack';
 import MainTab from './screens/main/mainTab';
-
 
 const Stack = createStackNavigator();
