--- assets/styles/styleVars.js
@@ -1,4 +1,0 @@
-export const colors = {
-    blue: '#42aaf5',
-    red: '#e33232',
-};
